// src/app/features/matches/models/football-event.model.ts
export interface FootballEvent {
  id: number;
  match_id: number;
  event_number?: number | null;
  event_qtr?: number | null;
  ball_on?: number | null;
  ball_moved_to?: number | null;
  ball_picked_on?: number | null;
  ball_picked_on_fumble?: number | null;
  ball_kicked_to?: number | null;
  ball_returned_to?: number | null;
  ball_returned_to_on_fumble?: number | null;
  distance_on_offence?: number | null;
  offense_team?: number | null;
  event_qb?: number | null;
  event_down?: number | null;
  event_distance?: number | null;
  event_strong_side?: string | null;
  event_hash?: string | null;
  play_direction?: string | null;
  play_type?: string | null;
  play_result?: string | null;
  score_result?: string | null;
  is_fumble?: boolean | null;
  is_fumble_recovered?: boolean | null;
  run_player?: number | null;
  pass_received_player?: number | null;
  pass_dropped_player?: number | null;
  pass_deflected_player?: number | null;
  pass_intercepted_player?: number | null;
  fumble_player?: number | null;
  fumble_recovered_player?: number | null;
  tackle_player?: number | null;
  assist_tackle_player?: number | null;
  sack_player?: number | null;
  score_player?: number | null;
  defence_score_player?: number | null;
  kick_player?: number | null;
  kickoff_player?: number | null;
  return_player?: number | null;
  pat_one_player?: number | null;
  flagged_player?: number | null;
  punt_player?: number | null;
}

export interface FootballEventCreate {
  match_id: number;
  event_number?: number | null;
  event_qtr?: number | null;
  ball_on?: number | null;
  ball_moved_to?: number | null;
  ball_picked_on?: number | null;
  ball_picked_on_fumble?: number | null;
  ball_kicked_to?: number | null;
  ball_returned_to?: number | null;
  ball_returned_to_on_fumble?: number | null;
  distance_on_offence?: number | null;
  offense_team?: number | null;
  event_qb?: number | null;
  event_down?: number | null;
  event_distance?: number | null;
  event_strong_side?: string | null;
  event_hash?: string | null;
  play_direction?: string | null;
  play_type?: string | null;
  play_result?: string | null;
  score_result?: string | null;
  is_fumble?: boolean | null;
  is_fumble_recovered?: boolean | null;
  run_player?: number | null;
  pass_received_player?: number | null;
  pass_dropped_player?: number | null;
  pass_deflected_player?: number | null;
  pass_intercepted_player?: number | null;
  fumble_player?: number | null;
  fumble_recovered_player?: number | null;
  tackle_player?: number | null;
  assist_tackle_player?: number | null;
  sack_player?: number | null;
  score_player?: number | null;
  defence_score_player?: number | null;
  kick_player?: number | null;
  kickoff_player?: number | null;
  return_player?: number | null;
  pat_one_player?: number | null;
  flagged_player?: number | null;
  punt_player?: number | null;
}

export interface FootballEventUpdate {
  event_number?: number | null;
  event_qtr?: number | null;
  ball_on?: number | null;
  ball_moved_to?: number | null;
  ball_picked_on?: number | null;
  ball_picked_on_fumble?: number | null;
  ball_kicked_to?: number | null;
  ball_returned_to?: number | null;
  ball_returned_to_on_fumble?: number | null;
  distance_on_offence?: number | null;
  offense_team?: number | null;
  event_qb?: number | null;
  event_down?: number | null;
  event_distance?: number | null;
  event_strong_side?: string | null;
  event_hash?: string | null;
  play_direction?: string | null;
  play_type?: string | null;
  play_result?: string | null;
  score_result?: string | null;
  is_fumble?: boolean | null;
  is_fumble_recovered?: boolean | null;
  run_player?: number | null;
  pass_received_player?: number | null;
  pass_dropped_player?: number | null;
  pass_deflected_player?: number | null;
  pass_intercepted_player?: number | null;
  fumble_player?: number | null;
  fumble_recovered_player?: number | null;
  tackle_player?: number | null;
  assist_tackle_player?: number | null;
  sack_player?: number | null;
  score_player?: number | null;
  defence_score_player?: number | null;
  kick_player?: number | null;
  kickoff_player?: number | null;
  return_player?: number | null;
  pat_one_player?: number | null;
  flagged_player?: number | null;
  punt_player?: number | null;
}